drm: Add private data field to trace control block
We need to be able to store private data in the tcb across it's
lifetime. To ensure proper destruction of the data a free_priv_data
callback must be provided if an allocation is stored in priv_data. The
callback is executed automatically when the life of the tcb ends.

* defs.h (get_tcb_priv_data, set_tcb_priv_data): extern declaration.
(struct tcb): Add _priv_data and _free_priv_data.
* strace.c (get_tcb_priv_data, set_tcb_priv_data)
(free_tcb_priv_data): New function
(drop_tcb): Execute free_tcb_priv_data callback
* syscall.c (trace_syscall_exiting): Execute free_tcb_priv_data callback

Signed-off-by: Patrik Jakobsson <patrik.jakobsson@linux.intel.com>